script: Move more files from `script/dom/`
It's increasingly difficult to figure out naming and locations
for these last few remaining files. These moves are my best
effort to find a reasonable place.

Part of #38901

Testing: it compiles
