+++ components/script/dom/mod.rs
@@ -215,6 +215,8 @@ pub(crate) use self::abort::*;
 pub(crate) mod activation;
 pub(crate) mod animations;
 pub(crate) use self::animations::*;
+pub(crate) mod attributes;
+pub(crate) use self::attributes::*;
 pub(crate) mod audio;
 pub(crate) use self::audio::*;
 pub(crate) mod bindings;
@@ -230,12 +232,11 @@ pub(crate) mod clipboard;
 pub(crate) use self::clipboard::*;
 pub(crate) mod console;
 pub(crate) mod cookiestore;
-pub(crate) mod cookiestoremanager;
+pub(crate) use self::cookiestore::*;
 pub(crate) mod credentialmanagement;
 pub(crate) use self::credentialmanagement::*;
 pub(crate) mod css;
 pub(crate) use self::css::*;
-pub(crate) mod customelementregistry;
 pub(crate) mod datatransfer;
 pub(crate) use self::datatransfer::*;
 pub(crate) mod debugger;
@@ -243,11 +244,6 @@ pub(crate) use self::debugger::*;
 pub(crate) mod document;
 pub(crate) use self::document::*;
 pub(crate) mod domexception;
-pub(crate) mod domimplementation;
-pub(crate) mod domparser;
-pub(crate) mod domstringlist;
-pub(crate) mod domstringmap;
-pub(crate) mod domtokenlist;
 pub(crate) mod element;
 pub(crate) use self::element::*;
 pub(crate) mod encoding;
@@ -366,9 +362,9 @@ pub(crate) mod workers;
 pub(crate) use self::workers::*;
 pub(crate) mod worklet;
 pub(crate) use self::worklet::*;
-pub(crate) mod xmldocument;
+pub(crate) mod xml;
+pub(crate) use self::xml::*;
 pub(crate) mod xmlhttprequest;
 pub(crate) use self::xmlhttprequest::*;
-pub(crate) mod xmlserializer;
 pub(crate) mod xpath;
 pub(crate) use self::xpath::*;
