+++ python/tidy/error_message_exceptions.txt
@@ -1,5 +1,7 @@
 ./components/script/dom/abort/abortsignal.rs 3
 ./components/script/dom/animations/keyframeeffect.rs 3
+./components/script/dom/attributes/domstringmap.rs 2
+./components/script/dom/attributes/domtokenlist.rs 4
 ./components/script/dom/audio/analysernode.rs 7
 ./components/script/dom/audio/audiobuffer.rs 7
 ./components/script/dom/audio/audiobuffersourcenode.rs 1
@@ -33,21 +35,18 @@
 ./components/script/dom/characterdata/text.rs 1
 ./components/script/dom/clipboard/clipboard.rs 1
 ./components/script/dom/clipboard/clipboarditem.rs 2
-./components/script/dom/cookiestore.rs 8
+./components/script/dom/cookiestore/cookiestore.rs 8
 ./components/script/dom/credentialmanagement/credentialscontainer.rs 8
 ./components/script/dom/credentialmanagement/passwordcredential.rs 1
 ./components/script/dom/css/css.rs 2
 ./components/script/dom/css/cssrulelist.rs 4
 ./components/script/dom/css/cssstyledeclaration.rs 3
 ./components/script/dom/css/fontface.rs 4
-./components/script/dom/customelementregistry.rs 7
 ./components/script/dom/datatransfer/datatransferitemlist.rs 1
 ./components/script/dom/document/document.rs 28
 ./components/script/dom/document/documentorshadowroot.rs 1
+./components/script/dom/document/domimplementation.rs 1
 ./components/script/dom/document/websocket.rs 7
-./components/script/dom/domimplementation.rs 1
-./components/script/dom/domstringmap.rs 2
-./components/script/dom/domtokenlist.rs 4
 ./components/script/dom/element/element.rs 11
 ./components/script/dom/element/namednodemap.rs 2
 ./components/script/dom/event/eventtarget.rs 1
@@ -74,6 +73,7 @@
 ./components/script/dom/intersectionobserver/intersectionobserver.rs 2
 ./components/script/dom/macros.rs 2
 ./components/script/dom/navigator/navigator.rs 3
+./components/script/dom/nodecustomelementregistry.rs 7
 ./components/script/dom/performance/performanceobserver.rs 4
 ./components/script/dom/range/range.rs 20
 ./components/script/dom/selection.rs 14
